Refactor to use real API hooks instead of mock API; update folder filtering logic to include child folders
